refactor: decouple cli (#19)
