+++ pyproject.toml
@@ -32,7 +32,7 @@ packages = [
 license = "Apache-2.0"
 
 [tool.poetry.scripts]
-encord-active = "encord_active.app.cli.main:cli"
+encord-active = "encord_active.cli.main:cli"
 
 [tool.poetry.dependencies]
 python = ">=3.9,<3.9.7 || >3.9.7,<3.11"
@@ -91,7 +91,7 @@ tuna = "^0.5.11"
 poethepoet = "^0.16.5"
 
 [tool.poe.tasks]
-importtime = { shell = "python -X importtime -c 'import encord_active.app.cli.main' 2> imports.prof && python -m tuna imports.prof" }
+importtime = { shell = "python -X importtime -c 'import encord_active.cli.main' 2> imports.prof && python -m tuna imports.prof" }
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
